endpoint: move api-related functions to `api.go`
Signed-off by: Ian Vernon <ian@cilium.io>
